# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoanFlow API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listMyLoans
      tags: [loans]
      summary: List my loan applications
      responses:
        "200":
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Submit a new loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/BadRequest"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loan detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /loans/{id}/upload-signed-contract:
    post:
      operationId: uploadSignedContract
      tags: [contracts]
      summary: User uploads the signed contract PDF (base64)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadFileInput"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /loans/{id}/withdraw:
    post:
      operationId: withdrawFunds
      tags: [wallet]
      summary: Withdraw available loan funds (simulation)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoanStatus:
      type: string
      enum:
        - EN_ATTENTE
        - ACCEPTE
        - REFUSE
        - CONTRAT_ENVOYE
        - CONTRAT_SIGNE
        - EN_TRAITEMENT
        - FONDS_DISPONIBLES
...
    UploadedDocument:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
        dataBase64:
          type: string
      required: [filename, contentType, dataBase64]
    UploadedDocumentRef:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        contentType:
          type: string
        uploadedAt:
          type: string
      required: [id, filename, contentType, uploadedAt]
    CreateLoanInput:
      type: object
      properties:
        applicantName:
          type: string
          minLength: 2
        applicantEmail:
          type: string
        amount:
          type: number
          minimum: 100
        durationMonths:
          type: integer
          minimum: 3
          maximum: 120
        monthlyIncome:
          type: number
          minimum: 0
        purpose:
          type: ["string", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/UploadedDocument"
      required: [applicantName, applicantEmail, amount, durationMonths, monthlyIncome, documents]
    Loan:
      type: object
      properties:
        id:
          type: string
        applicantName:
          type: string
        applicantEmail:
          type: string
        amount:
          type: number
        durationMonths:
          type: integer
        monthlyIncome:
          type: number
        purpose:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LoanStatus"
        decisionAt:
          type: ["string", "null"]
        contractSignedAt:
          type: ["string", "null"]
        fundsAvailableAt:
          type: ["string", "null"]
        processingUntil:
          type: ["string", "null"]
        withdrawnAmount:
          type: number
        availableBalance:
          type: number
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - applicantName
        - applicantEmail
        - amount
        - durationMonths
        - monthlyIncome
        - purpose
        - status
        - decisionAt
        - contractSignedAt
        - fundsAvailableAt
        - processingUntil
        - withdrawnAmount
        - availableBalance
        - adminNote
        - createdAt
        - updatedAt
    LoanDetail:
      allOf:
        - $ref: "#/components/schemas/Loan"
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: "#/components/schemas/UploadedDocumentRef"
            generatedContract:
              type: ["object", "null"]
              properties:
                id:
                  type: string
                filename:
                  type: string
                uploadedAt:
                  type: string
              required: [id, filename, uploadedAt]
            signedContract:
              type: ["object", "null"]
              properties:
                id:
                  type: string
                filename:
                  type: string
                uploadedAt:
                  type: string
              required: [id, filename, uploadedAt]
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
          required: [documents, generatedContract, signedContract, timeline]
...
    UploadFileInput:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
        dataBase64:
          type: string
      required: [filename, contentType, dataBase64]
    WithdrawInput:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
      required: [amount]
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required: [id, kind, message, createdAt]
...